requirements.txt
Pillow
pytesseract
python-dotenv
pyTelegramBotAPI